requirements.txt
retina-face
opencv-python
Pillow
numpy
tf-keras
# sqlite3 is included with Python standard library